#1548 util package
