+++ eo-maven-plugin/src/main/java/org/eolang/maven/DcsDefault.java
@@ -35,6 +35,7 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Optional;
 import org.apache.maven.model.Dependency;
+import org.eolang.maven.util.Rel;
 
 /**
  * It is a list of dependencies that are needed by the build.
