+++ eo-maven-plugin/src/main/java/org/eolang/maven/RegisterMojo.java
@@ -34,6 +34,8 @@ import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.cactoos.set.SetOf;
+import org.eolang.maven.util.Rel;
+import org.eolang.maven.util.Walk;
 
 /**
  * Find and register all {@code .eo} sources in the "foreign" catalog.
