+++ eo-maven-plugin/src/main/java/org/eolang/maven/SpyTrain.java
@@ -32,6 +32,7 @@ import com.yegor256.xsline.TrLambda;
 import com.yegor256.xsline.Train;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import org.eolang.maven.util.Home;
 
 /**
  * Train that spies.
