+++ eo-maven-plugin/src/main/java/org/eolang/maven/footprint/FtDefault.java
@@ -29,8 +29,8 @@ import org.cactoos.Scalar;
 import org.cactoos.scalar.IoChecked;
 import org.cactoos.text.IoCheckedText;
 import org.cactoos.text.TextOf;
-import org.eolang.maven.Home;
 import org.eolang.maven.Place;
+import org.eolang.maven.util.Home;
 
 /**
  * Default implementation of a Footprint.
