+++ eo-maven-plugin/src/main/java/org/eolang/maven/util/Home.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven;
+package org.eolang.maven.util;
 
 import com.jcabi.log.Logger;
 import java.io.IOException;
