+++ eo-maven-plugin/src/main/java/org/eolang/maven/util/package-info.java
@@ -21,49 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven;
-
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.Arrays;
-import org.cactoos.bytes.BytesOf;
-import org.cactoos.bytes.Md5DigestOf;
-import org.cactoos.bytes.UncheckedBytes;
-import org.cactoos.io.InputOf;
-
 /**
- * MD5 hash of a file (its content).
- *
- * @since 0.24
+ * Utility classes for EO maven plugin.
  */
-final class FileHash {
-
-    /**
-     * The file.
-     */
-    private final Path file;
-
-    /**
-     * Ctor.
-     * @param path The name of the file
-     */
-    FileHash(final Path path) {
-        this.file = path;
-    }
-
-    @Override
-    public String toString() {
-        final String hash;
-        if (Files.exists(this.file)) {
-            hash = Arrays.toString(
-                new UncheckedBytes(
-                    new Md5DigestOf(new InputOf(new BytesOf(this.file)))
-                ).asBytes()
-            );
-        } else {
-            hash = "";
-        }
-        return hash;
-    }
-
-}
+package org.eolang.maven.util;
