+++ eo-maven-plugin/src/test/java/org/eolang/maven/AssembleMojoTest.java
@@ -28,6 +28,7 @@ import java.nio.file.Paths;
 import org.cactoos.io.InputOf;
 import org.cactoos.set.SetOf;
 import org.eolang.maven.objectionary.Objectionary;
+import org.eolang.maven.util.Home;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
