+++ eo-maven-plugin/src/test/java/org/eolang/maven/DcsJsonTest.java
@@ -27,6 +27,7 @@ import java.io.IOException;
 import java.nio.file.Path;
 import org.cactoos.io.ResourceOf;
 import org.cactoos.scalar.LengthOf;
+import org.eolang.maven.util.Home;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.io.TempDir;
