+++ eo-maven-plugin/src/test/java/org/eolang/maven/FileHashTest.java
@@ -25,6 +25,8 @@ package org.eolang.maven;
 
 import java.io.IOException;
 import java.nio.file.Path;
+import org.eolang.maven.util.FileHash;
+import org.eolang.maven.util.Home;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
