+++ eo-maven-plugin/src/test/java/org/eolang/maven/PlaceMojoTest.java
@@ -27,6 +27,7 @@ import com.yegor256.tojos.Tojos;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import org.eolang.maven.util.Home;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
