+++ eo-maven-plugin/src/test/java/org/eolang/maven/SkipTest.java
@@ -29,6 +29,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import org.cactoos.io.InputOf;
 import org.eolang.maven.objectionary.Objectionary;
+import org.eolang.maven.util.Home;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
