+++ eo-maven-plugin/src/test/java/org/eolang/maven/optimization/OptCachedTest.java
@@ -32,7 +32,7 @@ import java.nio.file.Paths;
 import org.cactoos.bytes.BytesOf;
 import org.cactoos.bytes.UncheckedBytes;
 import org.cactoos.io.ResourceOf;
-import org.eolang.maven.Home;
+import org.eolang.maven.util.Home;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.hamcrest.io.FileMatchers;
